# Make compaction completion-based instead of readiness-based.