This is the first step of https://github.com/tigerbeetledb/tigerbeetle/issues/269. 

Compactions are still concurrent per tree but are uncoupled from the pace of beats. Instead, they start at the beginning of each half-bar, and at the end of each half-bar we wait for them to finish.

As a minor win, the merging itself can now be a really tight loop over large batches of values and has vanished from the perf profile.

```
# before
1223 batches in 94.13 s
load offered = 1000000 tx/s
load accepted = 106236 tx/s
batch latency p00 = 5 ms
batch latency p10 = 31 ms
batch latency p20 = 32 ms
batch latency p30 = 32 ms
batch latency p40 = 32 ms
batch latency p50 = 32 ms
batch latency p60 = 32 ms
batch latency p70 = 32 ms
batch latency p80 = 37 ms
batch latency p90 = 37 ms
batch latency p100 = 3885 ms
transfer latency p00 = 5 ms
transfer latency p10 = 3373 ms
transfer latency p20 = 9848 ms
transfer latency p30 = 16901 ms
transfer latency p40 = 25117 ms
transfer latency p50 = 34823 ms
transfer latency p60 = 44714 ms
transfer latency p70 = 54532 ms
transfer latency p80 = 64492 ms
transfer latency p90 = 74352 ms
transfer latency p100 = 84099 ms

# after
1223 batches in 93.21 s
load offered = 1000000 tx/s
load accepted = 107289 tx/s
batch latency p00 = 5 ms
batch latency p10 = 31 ms
batch latency p20 = 32 ms
batch latency p30 = 32 ms
batch latency p40 = 32 ms
batch latency p50 = 32 ms
batch latency p60 = 37 ms
batch latency p70 = 37 ms
batch latency p80 = 37 ms
batch latency p90 = 37 ms
batch latency p100 = 3690 ms
transfer latency p00 = 5 ms
transfer latency p10 = 3558 ms
transfer latency p20 = 9408 ms
transfer latency p30 = 16161 ms
transfer latency p40 = 24337 ms
transfer latency p50 = 33938 ms
transfer latency p60 = 43993 ms
transfer latency p70 = 54132 ms
transfer latency p80 = 64300 ms
transfer latency p90 = 74033 ms
transfer latency p100 = 83175 ms
```

I tried optionally bypassing the cache in `jamii-bypass-cache` but it's only a small win for extra grid complexity. Worth retrying after merging https://github.com/tigerbeetledb/tigerbeetle/pull/564 though.

Follow-up work:
* Make compactions serial.
* Queue multiple reads inside table and level iterators.
* Queue multiple writes inside compaction.